+++ adventure-beta/package.json
@@ -11,9 +11,11 @@
     "@types/node": "12.12.55",
     "@types/react": "16.9.49",
     "@types/react-dom": "16.9.8",
+    "@types/react-router-dom": "5.1.5",
     "graphql": "15.3.0",
     "react": "16.13.1",
     "react-dom": "16.13.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
     "typescript": "3.7.5"
   },
