+++ adventure-beta/src/container/BlankPage/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function BlankPage() {
+  return <h1>Pós Login</h1>;
+}
+
+export default BlankPage;
